Update README.md
I worked on a Resume Analyzer project that leverages Natural Language Processing (NLP) and Machine Learning to automate the process of analyzing job applicants' resumes. The primary objective was to evaluate the suitability of a candidate for a specific role by comparing their resume content with a predefined set of job requirements and skills.
For this project, I used the Sentence-Transformer model to analyze and compare the textual content of resumes against job descriptions. The model utilizes advanced NLP techniques like semantic similarity to measure how well a resume matches a given job role. I also used Python, Streamlit for the user interface, and libraries like PyMuPDF for extracting text from PDF resumes